requirements.txt
Flask>=3.0.0
PyYAML>=6.0.0
pandas>=2.0.0
matplotlib>=3.8.0
scikit-optimize>=0.10.2
